cantidad de materiales double

+++ models/models.py
@@ -9,7 +9,8 @@ class ProyectoMateriales(models.Model):
 
     name = fields.Char(string="Nombre", required=False, )
     product_id = fields.Many2one(comodel_name="product.template", string="Producto", required=False, )
-    product_qty = fields.Integer(string="Cantidad", required=False, )
+    product_qty = fields.Float(string="Cantidad",  required=False, )
+
     project_id = fields.Many2one(comodel_name="project.project", string="Projecto", required=False, )
 
 class NewModule(models.Model):
